backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
requests==2.31.0
beautifulsoup4==4.12.0
dnspython==2.4.2
tldextract==5.0.0
urllib3==2.1.0
httpx==0.25.0
